[FIX] FeedComment, Likes, BookMark, FeedDetailLocImg 각각 다른 컨트롤러와 서비스단으로 분리.
