+++ src/main/java/com/example/trip/controller/FeedController.java
@@ -44,38 +44,6 @@ public class FeedController {
                 .build(), HttpStatus.OK);
     }
 
-    @Caching(evict = {
-            @CacheEvict(value = "feedlist"),
-            @CacheEvict(value = "feed", key = "#feedId"),
-            @CacheEvict(value = "feeddetailloc", key = "#feeddetaillocId")
-    })
-    @PostMapping("/feed/image/{feedDetailLocId}")
-    public ResponseEntity<FeedResponseDto.FeedResponseOptional> registerFeedImage(
-            @PathVariable Long feedDetailLocId,
-            @RequestPart(value="imgFiles") List<MultipartFile> imgFiles) {
-        Map<String, String> nameAndUrl = feedService.registerFeedImage(feedDetailLocId, imgFiles);
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseOptional.builder()
-                .result("success")
-                .msg("피드 이미지 등록 성공하였습니다.")
-                .data(nameAndUrl)
-                .build(), HttpStatus.OK);
-    }
-
-    @Caching(evict = {
-            @CacheEvict(value = "feedlist"),
-            @CacheEvict(value = "feed", key = "#feedId"),
-            @CacheEvict(value = "feeddetailloc", key = "#feeddetaillocId")
-    })
-    @DeleteMapping("/feed/image")
-    public ResponseEntity<FeedResponseDto.FeedResponseDefault> deleteFeedImage(
-            @RequestBody FeedRequestDto.FeedRequestDeleteImgDto  FeedRequestDeleteImgDto) {
-        feedService.deleteFeedImage(FeedRequestDeleteImgDto);
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
-                .result("success")
-                .msg("피드 이미지 삭제 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
     @PutMapping("/feed/{feedId}")
     public ResponseEntity<FeedResponseDto.FeedResponseDefault> modifyFeed(
             @AuthenticationPrincipal UserDetailsImpl userDetails,
@@ -100,75 +68,6 @@ public class FeedController {
         return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
                 .result("success")
                 .msg("피드 삭제 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
-    @PostMapping("/feed/{feedDetailLocId}/like")
-    public ResponseEntity<FeedResponseDto.FeedResponseDefault> likeFeed(@PathVariable Long feedDetailLocId,
-                                                    @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        feedService.likeFeed(feedDetailLocId, userDetails.getUser());
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
-                .result("success")
-                .msg("피드 좋아요 등록에 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/feed/{feedDetailLocId}/unlike")
-    public ResponseEntity<FeedResponseDto.FeedResponseOptional> unlikeFeed(@PathVariable Long feedDetailLocId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        feedService.unlikeFeed(feedDetailLocId, userDetails.getUser());
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseOptional.builder()
-                .result("success")
-                .msg("피드 좋아요 취소에 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
-    @PostMapping("/feed/{feedDetailLocId}/bookmark")
-    public ResponseEntity<FeedResponseDto.FeedResponseDefault> bookmarkFeed(@PathVariable Long feedDetailLocId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        feedService.bookmarkFeed(feedDetailLocId, userDetails.getUser());
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
-                .result("success")
-                .msg("피드 북마크 등록에 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/feed/{feedDetailLocId}/unbookmark")
-    public ResponseEntity<FeedResponseDto.FeedResponseDefault> unbookmarkFeed(@PathVariable Long feedDetailLocId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        feedService.unbookmarkFeed(feedDetailLocId, userDetails.getUser());
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
-                .result("success")
-                .msg("피드 북마크 취소에 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
-    @PostMapping("/feed/comment/{feedDetailLocId}")
-    public ResponseEntity<FeedResponseDto.FeedResponseDefault> registerFeedComment(@AuthenticationPrincipal UserDetailsImpl userDetails,@PathVariable Long feedDetailLocId, @RequestBody FeedRequestDto.FeedRequestCommentRegisterDto feedRequestCommentRegisterDto) {
-        feedService.registerFeedComment(userDetails.getUser(), feedDetailLocId, feedRequestCommentRegisterDto);
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
-                .result("success")
-                .msg("댓글 등록 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
-    @PutMapping("/feed/comment/{commentId}")
-    public ResponseEntity<FeedResponseDto.FeedResponseDefault> modifyFeedComment(
-            @AuthenticationPrincipal UserDetailsImpl userDetails,
-            @PathVariable Long commentId,
-            @RequestBody FeedRequestDto.FeedRequestCommentModifyDto feedRequestCommentModifyDto) {
-        feedService.modifyFeedComment(userDetails.getUser(), commentId, feedRequestCommentModifyDto);
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
-                .result("success")
-                .msg("댓글 수정 성공하였습니다.")
-                .build(), HttpStatus.OK);
-    }
-
-    @DeleteMapping("/feed/comment/{commentId}")
-    public ResponseEntity<FeedResponseDto.FeedResponseDefault> deleteFeedComment(
-            @AuthenticationPrincipal UserDetailsImpl userDetails,
-            @PathVariable Long commentId) {
-        feedService.deleteFeedComment(userDetails.getUser(), commentId);
-        return new ResponseEntity<>(FeedResponseDto.FeedResponseDefault.builder()
-                .result("success")
-                .msg("댓글 삭제 성공하였습니다.")
                 .build(), HttpStatus.OK);
     }
 }
